ERC20 token updates
added framework libraries for ERC165 and ERC173, as well as basic ERC20 token support, as well as a custom, non-standard SafeERC20Token, which requires contracts to implement iERC20Receiver, if they wish to accept ERC20 transfers. The default behavior is ERC20 token transfer to a contract address are automatically accepted, causing the tokens to become lost forever if there is no mechanism to transfer those tokens from within the contract.

SafeERC20Token prevents this class of bug by automatically denying transfers to any contract which does NOT implement iERC20Receiver