ENS domain added to README files & ERC/EIP Encoder Factories
added official ENS domain name vigilance.eth to all README.md files.

included ERC and EIP Encoder factories, to allow for creation of singleton contracts which CREATE2 to have the same address on every network which they are deployed, meaning there should only ever be a single contract (or a single version of a contract with the same creation code) per blockchain.

As a security measure, solidarity EIP-2470 contracts implement ERC-173 ownership/EIP-801 Canary so that only the address which deploys the factory (or current owner, if ownership has transferred) may deploy additional contracts.

WARNING: Renouncing ownership of a singleton factory prevents any further attempts to deploy new contracts and is irreversible. Useful for developing, then renounce ownership when live.

These singleton factories can also deploy additional contracts which is either a new version or uses a different hash/salt, to allow the release of multiple versions of a contract, all of which can be accessed using the factory's deployedContracts method